Move methods to sql executor
