+++ src/mayim/executor/mysql.py
@@ -1,9 +1,8 @@
 from __future__ import annotations
 
-from contextlib import asynccontextmanager
+
 from typing import Any, Dict, Optional, Sequence
 
-from mayim.exception import MayimError
 from mayim.query.mysql import MysqlQuery
 
 from .sql import SQLExecutor
@@ -39,18 +38,5 @@ class MysqlExecutor(SQLExecutor):
                 raw = await getattr(cursor, method_name)()
                 return raw
 
-    def _get_method(self, as_list: bool):
-        return "fetchall" if as_list else "fetchone"
-
-    @asynccontextmanager
-    async def transaction(self):
-        async with self.pool.connection() as conn:
-            self.pool._connection.set(conn)
-            yield conn
-            self.pool._connection.set(None)
-
-    async def rollback(self) -> None:
-        existing = self.pool._connection.get(None)
-        if not existing:
-            raise MayimError("Cannot rollback non-existing transaction")
+    async def _rollback(self, _) -> None:
         self.pool._commit.set(False)  # type: ignore
