+++ src/mayim/executor/postgres.py
@@ -1,9 +1,7 @@
 from __future__ import annotations
 
-from contextlib import asynccontextmanager
-from typing import Any, Coroutine, Dict, Optional, Sequence
+from typing import Any, Dict, Optional, Sequence
 
-from mayim.exception import MayimError
 from mayim.query.postgres import PostgresQuery
 
 from .sql import SQLExecutor
@@ -39,18 +37,5 @@ class PostgresExecutor(SQLExecutor):
             raw = await getattr(cursor, method_name)()
             return raw
 
-    def _get_method(self, as_list: bool):
-        return "fetchall" if as_list else "fetchone"
-
-    @asynccontextmanager
-    async def transaction(self):
-        async with self.pool.connection() as conn:
-            self.pool._connection.set(conn)
-            yield conn.transaction()
-            self.pool._connection.set(None)
-
-    def rollback(self) -> Coroutine[Any, Any, Any]:
-        existing = self.pool._connection.get(None)
-        if not existing:
-            raise MayimError("Cannot rollback non-existing transaction")
-        return existing.rollback()
+    async def _rollback(self, existing) -> None:
+        await existing.rollback()
